requirements.txt
# Core dependencies
pyyaml>=6.0.1
websocket-client>=1.6.4
jsonschema>=4.20.0
Pillow>=10.1.0
python-dotenv>=1.0.0
pydantic>=2.5.2

# Testing dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
